+++ collections/src/includes/rchain.h
@@ -3,7 +3,7 @@
  */
 #ifndef RCHAIN_H_
 #define RCHAIN_H_
-#include"node.h"
+#include"node.cpp"
 namespace sktech{
 	template<class T>
 	class rchain{
@@ -20,20 +20,20 @@ namespace sktech{
 		const T &back()const;
 		void pop_front();
 		void pop_back();
-		const size_t size()const;
-		const size_t Max()const;
+		const unsigned long size()const;
+		const unsigned long Max()const;
 		bool full()const;
 		bool empty()const;
 		rchain<T> &operator=(const rchain<T> &otherChain)
-		T &operator[](unsigned n);
+		T &operator[](unsigned long n);
 	private:
 		bool grow(unsigned _size, const rchain<T> &otherChain = NULL, unsigned index = 0);
-		T &left(node<T> *nextNode, unsigned i, unsigned index);
-		T &right(node<T> *nextNode, unsigned i, unsigned index);
+		T &left(node<T> *nextNode, unsigned long i, unsigned long index);
+		T &right(node<T> *nextNode, unsigned long i, unsigned long index);
 		bool add(node<T> *addPoint, const T &newVal, bool first);
 		bool varAdd(va_list &newChain, unsigned new_size, unsigned index = 0);
 		bool remove(node<T> *deletePoint, bool first);
-		size_t _size, max;
+		unsigned long _size, max;
 		node<T> *firstP, *lastP;
 	};
 }
